fix: use correct merge API endpoint in LFID enrichment
